Rewrited VertexArray

+++ ui_opengl/math.cpp
@@ -10,8 +10,8 @@
 
 float getRotAngle(const V2f& a, const V2f& b) {
   V2f diff(
-      pow(b.x() - a.x(), 2),
-      pow(b.y() - a.y(), 2));
+      static_cast<float>(pow(b.x() - a.x(), 2)),
+      static_cast<float>(pow(b.y() - a.y(), 2)));
   float distance = sqrt(diff.x() + diff.y());
   float angle = rad2deg(asin((b.x() - a.x()) / distance));
   if (b.y() - a.y() > 0) {
